Picking and dropping objects. Physics: boxes replaced by spheres.
First prototype with using spheres for finger tips, enabling
picking and dropping boxes. Drop happens when finger dissapears
or when distance is larger than on pick.

A lot of noise to process in next patch, picled objects tend to
shake and jump around as fingertips are not stabile.

TODO: ammo.wasm version.

Updated screenshot animation.